refactor(files): satisfy strict frontend-structure checks for the tree overlay
- FileTreeRow: use the ui Button primitive instead of a raw <button>
  (unstyled variant, matching PaneFileTree's precedent)
- move file-tree-icon-colors.ts to lib/domain/files (components/** is
  .tsx-only)
- move use-tree-panel-resize.ts to hooks/ui/layout alongside the other
  resize hooks
- extract FileSearchResultsTree.tsx from FileTreeOverlay.tsx to get the
  overlay under the ~400-line strict threshold (438 -> 352)
