+++ apps/desktop/src/components/workspace/files/tree/FileTreeRow.tsx
@@ -1,7 +1,8 @@
 import { twMerge } from "@proliferate/ui/utils/tw-merge";
+import { Button } from "@proliferate/ui/primitives/Button";
 import { ChevronRight } from "@proliferate/ui/icons";
 import { FileTreeEntryIcon } from "@/components/workspace/files/file-icons";
-import { fileTreeIconToneClass } from "@/components/workspace/files/tree/file-tree-icon-colors";
+import { fileTreeIconToneClass } from "@/lib/domain/files/file-tree-icon-colors";
 
 interface FileTreeRowProps {
   name: string;
@@ -29,8 +30,10 @@ export function FileTreeRow({
   const iconTone = fileTreeIconToneClass(name, path, kind);
 
   return (
-    <button
+    <Button
       type="button"
+      variant="unstyled"
+      size="unstyled"
       role="treeitem"
       aria-expanded={isDirectory ? expanded : undefined}
       aria-selected={selected}
@@ -70,6 +73,6 @@ export function FileTreeRow({
           aria-label="Modified"
         />
       )}
-    </button>
+    </Button>
   );
 }
