+++ apps/desktop/src/lib/domain/files/file-tree-icon-colors.ts
